default S-exprssion indentation and expression navigation
Extracts the implementation of S-expression indentation and navigation
(forward, backward, up, and down) to a set of functions outside of
`racket:text-mixin`. That way, the implementation can be used in other
editing environments.

The `classified-text<%>` interface describes the methods that an
indentation or navigation implementation may call, but an object
is not required to explicitly implement `classified-text<%>`.